ocr-api/requirements.txt
# FastAPI
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# OCR
yomitoku>=0.11.0

# Image processing
pillow>=10.0.0
